Refactoring javascript.py
For continuity, the variable holding javascript code withing a
javascript object will be called statements and will be a list of either
strings or other javascript objects.

I would like to make different javascript objects inherit from eachother
(this seems like the right thing to do for continuity), but that'll have
to wait for the next commit!