Fixes issue mikedewar/d3py#23
This fixes the problem by allowing users to go into non-interactive mode
(endabled by default).  One problem this fix has made apparent is that
__del__ is not what we want to cleanup files: the reference count could
be at zero even if we have stale files lying around!

One solution would be to use __enter__ and __exit__ and require the user
to create a figure object with the `with` statement:

with Figure() as fig:
    ...do stuff with fig...

ideas?